DM working some people

+++ server/db/Tweet.js
@@ -36,7 +36,7 @@ const Tweet = db.define('tweets', {
         unique: true,
     },
     twitterUserId: {
-        type: db.Sequelize.INTEGER,
+        type: db.Sequelize.BIGINT,
     },
 });
